clinker-core: extract Source, Transform, Output, Composition arms to focused modules
Move the final four PlanNode arm bodies out of executor/dispatch.rs into
new sibling modules:

- source_dispatch.rs      — file/channel ingest, schema canonicalization,
                            on_unmapped policy, per-source rate counters,
                            node-rooted window finalize at EOF.
- transform_dispatch.rs    — record-level CXL projection / filter / distinct
                            / emit_each fan-out, plus the streaming-fused
                            fast path off a Source receiver.
- output_dispatch.rs       — writer init, schema mapping, include_unmapped
                            passthrough, source-file fan-out (emit_fan_out
                            moves here), correlation-buffer capture, and the
                            streaming-fused output short-circuit.
- composition_dispatch.rs  — body node-table swap, nested execute, and port
                            forwarding (collect_port_records and
                            execute_composition_body move here).

dispatch_plan_node now reads as a flat list of ten single delegating calls;
dispatch.rs keeps only the topo walker, the cross-arm shared helpers
(ExecutorContext, push_dlq, advance_cursor, check_dlq_rate, the node_buffer
ownership-transfer helpers, the fused-source consume/interleave paths) and
the correlation buffer entry types that ExecutorContext exposes.

Pure relocation: no symbol renames, no public-API change.

Closes #248, #109
