+++ crates/clinker-core/src/executor/dispatch.rs
@@ -22,7 +22,6 @@ use std::sync::{Arc, LazyLock};
 
 use clinker_record::{FieldMetadata, GroupByKey, PipelineCounters, Record, SchemaBuilder, Value};
 use cxl::eval::{EvalContext, ProgramEvaluator, SkipReason, StableEvalContext};
-use indexmap::IndexMap;
 use petgraph::Direction;
 use petgraph::graph::NodeIndex;
 
@@ -412,13 +411,11 @@ pub(crate) fn dispatch_transform_eval_error(
 use crate::executor::node_buffer::NodeBuffer;
 use crate::executor::schema_check::check_input_schema;
 use crate::executor::{
-    CompiledRoute, CompiledTransform, DlqEntry, build_format_writer, evaluate_single_transform,
-    evaluate_single_transform_windowed, stage_metrics,
+    CompiledRoute, CompiledTransform, DlqEntry, evaluate_single_transform, stage_metrics,
 };
 use crate::pipeline::memory::BudgetCategory;
 use crate::plan::bind_schema::CompileArtifacts;
 use crate::plan::execution::{ExecutionPlanDag, PlanNode};
-use crate::projection::project_output_from_record;
 use clinker_record::Schema;
 
 /// Mutable per-walk and borrowed plan-time state passed to
@@ -1065,110 +1062,6 @@ impl ExecutorContext<'_> {
                 .sum();
             self.source_count_per_source
                 .insert(Arc::clone(merged_key), Some(total));
-        }
-    }
-}
-
-/// Project every record in `rows` onto the column set in
-/// `buffer_schema`, returning narrow `Record` instances on a fresh,
-/// per-call `Arc<Schema>` shared by every emitted Record.
-///
-/// Mirrors the column-pruning pattern at
-/// `pipeline::arena::project_records_into_minimal`. Invoked once at the
-/// deferred-region producer (a relaxed-CK Aggregate) so the emit buffer
-/// retains exactly the columns the deferred operators reach via
-/// `Expr::support_into`. Source row numbers carry through unchanged.
-///
-/// `pub(crate)` so the commit-time `recompute_aggregates` phase can
-/// Emit a buffered record stream to a fan-out output: one writer per
-/// source-file `Arc<str>`, route each record to the writer keyed by
-/// its `$source.file` Arc. Writers without any matched records still
-/// flush an empty file (preserving header and any per-file framing).
-///
-/// All errors land in `output_errors` rather than short-circuiting so
-/// sibling writers in the same Output still get their chance to flush
-/// or report.
-#[allow(clippy::too_many_arguments)]
-fn emit_fan_out(
-    name: &str,
-    out_cfg: &crate::config::OutputConfig,
-    cxl_emit_names_opt: Option<&[String]>,
-    output_schema: &Arc<clinker_record::Schema>,
-    unbuffered: &[(Record, u64)],
-    per_file: HashMap<Arc<str>, Box<dyn Write + Send>>,
-    output_errors: &mut Vec<PipelineError>,
-    write_timer: &mut crate::executor::stage_metrics::CumulativeTimer,
-    projection_timer: &mut crate::executor::stage_metrics::CumulativeTimer,
-    collector: &mut crate::executor::stage_metrics::StageCollector,
-    scan_timer: crate::executor::stage_metrics::StageTimer,
-) {
-    use std::collections::HashMap as Hm;
-
-    // Build one format writer per pre-opened raw writer. Failed
-    // construction for one file does NOT abort the whole output —
-    // siblings still get their chance.
-    let mut format_writers: Hm<Arc<str>, Box<dyn clinker_format::FormatWriter>> = Hm::new();
-    for (file_arc, raw) in per_file {
-        match build_format_writer(out_cfg, raw, Arc::clone(output_schema)) {
-            Ok(fw) => {
-                format_writers.insert(file_arc, fw);
-            }
-            Err(e) => output_errors.push(e),
-        }
-    }
-    collector.record(scan_timer.finish(1, 1));
-
-    for (record, rn) in unbuffered {
-        let Some(file_path) = source_file_path_of(record) else {
-            output_errors.push(PipelineError::Internal {
-                op: "fan_out",
-                node: name.to_string(),
-                detail: format!(
-                    "row {rn} has no `$source.file` stamp; fan-out output requires per-record source-file lineage",
-                ),
-            });
-            continue;
-        };
-        // Look up the writer by path; the registry keys by Arc<str>
-        // so we need to find by string equality. Build a probing Arc
-        // once per record (cheap relative to the write itself).
-        let file_arc: Arc<str> = Arc::from(file_path);
-        let Some(fw) = format_writers.get_mut(&file_arc) else {
-            // Record's file isn't in the fan-out registry — typically
-            // means the CLI's writer setup didn't pre-open one for
-            // this file. Surface but keep going.
-            output_errors.push(PipelineError::Internal {
-                op: "fan_out",
-                node: name.to_string(),
-                detail: format!(
-                    "no fan-out writer registered for source file {:?}",
-                    file_arc
-                ),
-            });
-            continue;
-        };
-        let projected = {
-            let _guard = projection_timer.guard();
-            project_output_from_record(record, out_cfg, cxl_emit_names_opt)
-        };
-        let write_result = {
-            let _guard = write_timer.guard();
-            fw.write_record(&projected)
-        };
-        if let Err(e) = write_result {
-            output_errors.push(PipelineError::from(e));
-        }
-    }
-
-    // Flush every writer regardless of per-record errors so partial
-    // outputs land on disk for inspection.
-    for (_arc, mut fw) in format_writers {
-        let flush_result = {
-            let _guard = write_timer.guard();
-            fw.flush()
-        };
-        if let Err(e) = flush_result {
-            output_errors.push(PipelineError::from(e));
         }
     }
 }
@@ -2491,439 +2384,17 @@ pub(crate) fn dispatch_plan_node(
     }
     let node = current_dag.graph[node_idx].clone();
     match node {
-        PlanNode::Source { ref name, .. } => {
-            // Three input paths feed a Source's emit:
-            //
-            // 1. Source name in `ctx.fused_sources` — a downstream
-            //    `Merge.interleave` arm has taken ownership of this
-            //    Source's crossbeam `Receiver` and is consuming records
-            //    directly via `crossbeam_channel::Select`. This arm
-            //    returns cleanly without emitting; the fused Merge
-            //    populates the merge node's buffer.
-            // 2. Records already seeded into `ctx.node_buffers[node_idx]`
-            //    by the body executor at composition entry —
-            //    composition input ports surface as synthetic Source
-            //    nodes owning the records the parent scope harvested.
-            // 3. The live crossbeam `Receiver` in `source_records[name]`,
-            //    consumed via `recv` per record until the paired ingest
-            //    thread drops its sender. Per record: canonicalize onto
-            //    the source's plan-time schema, seed `$record.<key>`
-            //    defaults, seed `$source.<key>` defaults per
-            //    `(source, file_arc)`, advance the per-source running
-            //    counter. On `recv` returning `Err` (channel
-            //    disconnected), stamp the finalized per-source count and
-            //    call `finalize_node_rooted_windows` so every spec rooted
-            //    at this Source's `NodeIndex` lands its arena.
-            //
-            // Records are canonicalized onto the Source's plan-time
-            // `Arc<Schema>` so every downstream operator hits the
-            // `Arc::ptr_eq` fast path on the first record. Structural
-            // equality holds by construction.
-            if ctx.fused_sources.contains(name.as_str()) {
-                return Ok(());
-            }
-            let source_schema = current_dag.graph[node_idx].stored_output_schema().cloned();
-            let engine_stamped: Vec<(usize, Box<str>)> = source_schema
-                .as_ref()
-                .map(build_engine_stamped_tail)
-                .unwrap_or_default();
-            let canonicalize = |r: &Record| -> Record {
-                match source_schema.as_ref() {
-                    Some(target) => canonicalize_to_source_schema(r, target, &engine_stamped),
-                    None => r.clone(),
-                }
-            };
-
-            let (records, source_puncts): (
-                Vec<(Record, u64)>,
-                Vec<crate::executor::stream_event::Punctuation>,
-            ) = if let Some(seeded) = drain_node_buffer_slot(ctx, node_idx) {
-                // Body-context port source — records were seeded by
-                // `execute_composition_body` from parent-scope
-                // output. The seeded records still carry the parent
-                // producer's `Arc<Schema>`, so canonicalize them
-                // onto this port source's schema before downstream
-                // consumers run. Apply per-record seeding for body-
-                // declared record_vars (parent's writes survive via
-                // `seed_record_vars`'s preserve-existing semantics).
-                // Punctuations on the seeded port forward through to
-                // the Source's output buffer so downstream operators
-                // see the original document boundaries.
-                let mut out_records: Vec<(Record, u64)> = Vec::with_capacity(seeded.len_hint());
-                let mut out_puncts: Vec<crate::executor::stream_event::Punctuation> = Vec::new();
-                let has_record_seed = !ctx.record_var_seed.is_empty();
-                for event in seeded.drain() {
-                    match event? {
-                        crate::executor::stream_event::StreamEvent::Record(r, rn) => {
-                            let mut rec = canonicalize(&r);
-                            if has_record_seed {
-                                rec.seed_record_vars(ctx.record_var_seed);
-                            }
-                            seed_source_vars_for_record(ctx, name.as_str(), &rec)?;
-                            out_records.push((rec, rn));
-                        }
-                        crate::executor::stream_event::StreamEvent::Punctuation(p) => {
-                            out_puncts.push(p);
-                        }
-                    }
-                }
-                (out_records, out_puncts)
-            } else if let Some(rx) = ctx.source_records.remove(name.as_str()) {
-                // Live channel: consume per record so back-pressure
-                // engages — a slow upstream Source no longer blocks
-                // peers' channels from filling, and watermark
-                // observations on a different source's records
-                // flow through the dispatcher in the meantime
-                // wherever the executor task scheduler interleaves.
-                let timeout = ctx.idle_timeouts.get(name.as_str()).copied();
-                let has_record_seed = !ctx.record_var_seed.is_empty();
-                let source_name_arc: Arc<str> = Arc::from(name.as_str());
-                let mut drained: Vec<(Record, u64)> = Vec::new();
-                let mut drained_puncts: Vec<crate::executor::stream_event::Punctuation> =
-                    Vec::new();
-                // Tracked so an idle-timeout flips THAT file's
-                // partition to `Idle`. Before any record arrives the
-                // consumer uses the synthetic [`MERGED_SOURCE_FILE`]
-                // Arc, matching the engine-stamp path for record-
-                // less source contexts.
-                let mut last_file: Arc<str> = Arc::clone(&MERGED_SOURCE_FILE);
-                let mut count: u64 = 0;
-                let mut records_since_check: u32 = 0;
-                loop {
-                    let item: Option<crate::executor::stream_event::StreamEvent> = match timeout {
-                        Some(t) => match rx.recv_timeout(t) {
-                            Ok(item) => Some(item),
-                            Err(crossbeam_channel::RecvTimeoutError::Timeout) => {
-                                // Quiet for longer than
-                                // `idle_timeout` — flip the
-                                // partition tracked by `last_file`
-                                // to `Idle`. The next record un-
-                                // idles via `observe`. Idempotent on
-                                // repeat timeouts.
-                                ctx.watermarks.mark_idle(name.as_str(), &last_file);
-                                continue;
-                            }
-                            Err(crossbeam_channel::RecvTimeoutError::Disconnected) => None,
-                        },
-                        None => rx.recv().ok(),
-                    };
-                    match item {
-                        Some(crate::executor::stream_event::StreamEvent::Record(record, rn)) => {
-                            last_file = source_file_arc_of(&record);
-                            let mut rec = canonicalize(&record);
-                            if has_record_seed {
-                                rec.seed_record_vars(ctx.record_var_seed);
-                            }
-                            seed_source_vars_for_record(ctx, name.as_str(), &rec)?;
-                            if let Some(slot) = ctx.total_per_source.get_mut(&source_name_arc) {
-                                *slot += 1;
-                            }
-                            count += 1;
-                            records_since_check += 1;
-                            if records_since_check >= 1024 {
-                                records_since_check = 0;
-                                ctx.check_shutdown()?;
-                            }
-                            drained.push((rec, rn));
-                        }
-                        Some(crate::executor::stream_event::StreamEvent::Punctuation(p)) => {
-                            drained_puncts.push(p);
-                        }
-                        None => break,
-                    }
-                }
-                ctx.finalize_source_count(&source_name_arc, count);
-                if timeout.is_some() && ctx.watermarks.is_idle(name.as_str()) {
-                    tracing::debug!(
-                        target: "clinker::watermark",
-                        source = %name,
-                        "source consumer ended with all partitions in WatermarkStatus::Idle"
-                    );
-                }
-                (drained, drained_puncts)
-            } else {
-                // Defense-in-depth: a Source reaching this arm with
-                // neither a body-port seed, nor an entry in
-                // `source_records`, nor fused-bit set means the
-                // executor's unified ingest pass missed it. Every
-                // declared Source ingests through the same
-                // `source_records` map (no primary fallthrough), so
-                // any miss surfaces here as a loud internal error
-                // rather than the silent-corruption surface at the
-                // root of #47.
-                return Err(PipelineError::Internal {
-                    op: "executor",
-                    node: name.clone(),
-                    detail: format!(
-                        "Source '{name}' has no ingested records; \
-                         the executor's source-ingest pass missed this Source — \
-                         likely a planner regression introducing a Source topology \
-                         the ingest pass doesn't enumerate.",
-                    ),
-                });
-            };
-            // Build node-rooted arenas anchored at this Source's
-            // `NodeIndex`. Replaces the prologue's Phase-0 build:
-            // every spec previously rooted at `PlanIndexRoot::Source`
-            // now anchors here.
-            finalize_node_rooted_windows(ctx, current_dag, node_idx, &records)?;
-            tee_emit_to_region_input_buffers(ctx, current_dag, node_idx, &records)?;
-            let nb = admit_node_buffer(
-                ctx,
-                name,
-                node_idx,
-                records,
-                source_puncts,
-                node_buffer_spill_allowed(current_dag, node_idx),
-            )?;
-            ctx.node_buffers.insert(node_idx, nb);
+        PlanNode::Source { .. } => {
+            crate::executor::source_dispatch::dispatch_source(ctx, current_dag, node_idx, &node)?;
         }
 
-        PlanNode::Transform {
-            ref name,
-            window_index,
-            ..
-        } => {
-            // Streaming-fused path: when the pre-pass has flagged this
-            // Transform as eligible (sole upstream is a Source whose
-            // receiver lives in `ctx.source_records`, non-windowed,
-            // non-init-phase, no upstream fan-out, and the Source is
-            // not already claimed by Merge.interleave fusion), drive
-            // per-record evaluation directly off the receiver instead
-            // of consuming a Vec from `node_buffers`. See
-            // https://github.com/rustpunk/clinker/issues/74.
-            if ctx.should_fuse_transform(node_idx) {
-                return transform_fused_consume(ctx, current_dag, node_idx, name);
-            }
-            // Get input events: first check own buffer (set by Route
-            // node for branch dispatch), then fall back to predecessor.
-            // Records flow into per-record evaluation; punctuations are
-            // preserved verbatim and forwarded onto the output buffer
-            // alongside the transformed records (Preserving behavior).
-            let (input_records, input_puncts): (
-                Vec<(Record, u64)>,
-                Vec<crate::executor::stream_event::Punctuation>,
-            ) = if let Some(own_buf) = drain_node_buffer_slot(ctx, node_idx) {
-                own_buf.drain_split()?
-            } else {
-                let predecessors: Vec<NodeIndex> = current_dag
-                    .graph
-                    .neighbors_directed(node_idx, Direction::Incoming)
-                    .collect();
-                let pred_buf = if predecessors.len() == 1 {
-                    drain_node_buffer_slot(ctx, predecessors[0])
-                } else {
-                    predecessors
-                        .iter()
-                        .find_map(|p| drain_node_buffer_slot(ctx, *p))
-                };
-                match pred_buf {
-                    Some(nb) => nb.drain_split()?,
-                    None => (Vec::new(), Vec::new()),
-                }
-            };
-
-            // Find the CompiledTransform for this node
-            let transform_idx = match ctx.transform_by_name.get(name.as_str()) {
-                Some(&idx) => idx,
-                None => {
-                    // No transform found — pass through
-                    tee_emit_to_region_input_buffers(ctx, current_dag, node_idx, &input_records)?;
-                    let nb = admit_node_buffer(
-                        ctx,
-                        name.as_str(),
-                        node_idx,
-                        input_records,
-                        input_puncts,
-                        node_buffer_spill_allowed(current_dag, node_idx),
-                    )?;
-                    ctx.node_buffers.insert(node_idx, nb);
-                    return Ok(());
-                }
-            };
-
-            let mut evaluator = ProgramEvaluator::with_max_expansion(
-                Arc::clone(&ctx.compiled_transforms[transform_idx].typed),
-                ctx.compiled_transforms[transform_idx].has_distinct(),
-                ctx.compiled_transforms[transform_idx].max_expansion,
-            );
-
-            let expected_input = current_dag.graph[node_idx]
-                .expected_input_schema_in(current_dag)
-                .cloned();
-            let output_schema = current_dag.graph[node_idx].stored_output_schema().cloned();
-            let upstream_name = current_dag
-                .graph
-                .neighbors_directed(node_idx, Direction::Incoming)
-                .next()
-                .map(|i| current_dag.graph[i].name().to_string())
-                .unwrap_or_default();
-
-            // Plan invariant: any Transform with `window_index: Some`
-            // requires its WindowRuntime slot populated by either the
-            // source-rooted Phase-0 build (Source root) or the upstream
-            // operator's `finalize_node_rooted_windows` call (Node /
-            // ParentNode roots). Fail loudly if the slot is missing —
-            // the alternative (silently fall back to no-window eval)
-            // corrupts `$window.*` results.
-            if let Some(idx_num) = window_index {
-                let spec = current_dag.indices_to_build.get(idx_num).ok_or_else(|| {
-                    PipelineError::Internal {
-                        op: "executor",
-                        node: name.clone(),
-                        detail: format!(
-                            "transform {name:?} declares window_index {idx_num} \
-                             but plan.indices_to_build is too short"
-                        ),
-                    }
-                })?;
-                if ctx.window_runtime.resolve(&spec.root, idx_num).is_none() {
-                    return Err(PipelineError::Internal {
-                        op: "executor",
-                        node: name.clone(),
-                        detail: format!(
-                            "transform {name:?} declares window_index {idx_num} \
-                             but the runtime registry has no populated slot for that root; \
-                             upstream operator did not finalize its node-rooted windows"
-                        ),
-                    });
-                }
-            }
-
-            let mut output_records = Vec::with_capacity(input_records.len());
-
-            for (i, (record, rn)) in input_records.into_iter().enumerate() {
-                // Poll the shutdown flag every 1024 records so a long
-                // Transform chain terminates promptly on SIGINT without
-                // paying the atomic load per record.
-                if i > 0 && i.is_multiple_of(1024) {
-                    ctx.check_shutdown()?;
-                }
-                if let Some(exp) = expected_input.as_ref() {
-                    check_input_schema(exp, record.schema(), name, "transform", &upstream_name)?;
-                }
-                let source_file_arc = source_file_arc_of(&record);
-                let source_name_arc = source_name_arc_of(&record);
-                let eval_ctx = EvalContext {
-                    stable: ctx.stable,
-                    source_file: &source_file_arc,
-                    source_row: rn,
-                    source_path: &source_file_arc,
-                    source_count: ctx.source_count_by_name(&source_name_arc),
-                    source_batch: ctx.source_batch_arc,
-                    ingestion_timestamp: ctx.source_ingestion_timestamp,
-                    source_name: &source_name_arc,
-                    doc_ctx: record.doc_ctx(),
-                };
-
-                let target_schema = output_schema
-                    .as_ref()
-                    .cloned()
-                    .unwrap_or_else(|| Arc::clone(record.schema()));
-                let eval_result = {
-                    let _guard = ctx.transform_timer.guard();
-                    if let Some(idx_num) = window_index {
-                        // Resolve the WindowRuntime via the spec's root.
-                        // Source-rooted windows root at Phase-0's source
-                        // arena; node-rooted windows root at the
-                        // upstream operator's finalize. The plan-time
-                        // invariant check above guarantees `resolve`
-                        // returns `Some`; reaching `None` here is an
-                        // upstream-arm bug (e.g. forgetting to call
-                        // `finalize_node_rooted_windows` after emit).
-                        let spec = &current_dag.indices_to_build[idx_num];
-                        let runtime =
-                            ctx.window_runtime
-                                .resolve(&spec.root, idx_num)
-                                .ok_or_else(|| PipelineError::Internal {
-                                    op: "executor",
-                                    node: name.clone(),
-                                    detail: format!(
-                                        "transform {name:?} window_index {idx_num} \
-                                         resolves to no runtime at per-record dispatch; \
-                                         upstream finalize was skipped"
-                                    ),
-                                })?;
-                        // record_pos: enumerate index `i`. Every arena
-                        // is node-rooted; it was built from the
-                        // upstream's emit buffer in iteration order,
-                        // and the per-record dispatch loop iterates
-                        // the same buffer, so `i` equals the row's
-                        // arena position by construction.
-                        let record_pos = i as u64;
-                        evaluate_single_transform_windowed(
-                            &record,
-                            name,
-                            &mut evaluator,
-                            &eval_ctx,
-                            current_dag,
-                            idx_num,
-                            runtime,
-                            record_pos,
-                        )
-                    } else {
-                        evaluate_single_transform(
-                            &record,
-                            name,
-                            &mut evaluator,
-                            &eval_ctx,
-                            &target_schema,
-                        )
-                    }
-                };
-                match eval_result {
-                    Ok(records) => {
-                        if records.is_empty() {
-                            advance_cursor(ctx, &source_name_arc_of(&record), rn);
-                        } else {
-                            let mut emitted_any = false;
-                            for (modified_record, status) in records {
-                                match status {
-                                    Ok(()) => {
-                                        let advance_source = source_name_arc_of(&modified_record);
-                                        output_records.push((modified_record, rn));
-                                        advance_cursor(ctx, &advance_source, rn);
-                                        emitted_any = true;
-                                    }
-                                    Err(SkipReason::Filtered) => {
-                                        ctx.counters.filtered_count += 1;
-                                    }
-                                    Err(SkipReason::Duplicate) => {
-                                        ctx.counters.distinct_count += 1;
-                                    }
-                                }
-                            }
-                            if !emitted_any {
-                                advance_cursor(ctx, &source_name_arc_of(&record), rn);
-                            }
-                        }
-                    }
-                    Err((transform_name, eval_err)) => {
-                        dispatch_transform_eval_error(ctx, record, rn, transform_name, eval_err)?;
-                    }
-                }
-            }
-
-            // Materialize node-rooted window runtimes for any IndexSpec
-            // rooted at THIS Transform. The schema-changing case is the
-            // only one that drives node-rooted lookups here:
-            // pure-passthrough Transforms feed downstream windows
-            // through their predecessor's runtime (Sort/Route walk-
-            // through at lowering time covers passthroughs that did not
-            // change schema). When the lowering pass roots a window at
-            // this Transform's `NodeIndex`, the call below installs the
-            // matching runtime; otherwise the helper is a no-op.
-            finalize_node_rooted_windows(ctx, current_dag, node_idx, &output_records)?;
-            tee_emit_to_region_input_buffers(ctx, current_dag, node_idx, &output_records)?;
-            let nb = admit_node_buffer(
+        PlanNode::Transform { .. } => {
+            crate::executor::transform_dispatch::dispatch_transform(
                 ctx,
-                name,
+                current_dag,
                 node_idx,
-                output_records,
-                input_puncts,
-                node_buffer_spill_allowed(current_dag, node_idx),
+                &node,
             )?;
-            ctx.node_buffers.insert(node_idx, nb);
         }
 
         PlanNode::Route { .. } => {
@@ -2947,384 +2418,17 @@ pub(crate) fn dispatch_plan_node(
             )?;
         }
 
-        PlanNode::Output { ref name, .. } => {
-            // Streaming-Output short-circuit (issue #72). The executor
-            // entry already moved this Output's writer into a
-            // `std::thread` that drained records from a bounded crossbeam
-            // channel populated by the fused Merge arm. Per-record
-            // `write_record` already fired concurrently with Merge
-            // production; the dispatcher's end-of-DAG join surface joins
-            // the thread and folds its counters / timers / errors into
-            // the context. The Output's topo turn here is a no-op.
-            if ctx.streaming_output_nodes.contains(&node_idx) {
-                return Ok(());
-            }
-            // Get input records: check own buffer first (Route
-            // nodes store records at the successor's index), then
-            // fall back to predecessor buffers.
-            //
-            // Output is terminal — it writes to disk, so punctuations
-            // are consumed at this stage rather than forwarded. The
-            // input drain still uses `drain_split` for symmetry with
-            // every other operator, but the puncts vector goes unused
-            // for non-streaming outputs. Streaming Outputs (#72) take
-            // the early-return path above and forward puncts through
-            // the streaming channel separately.
-            let input_records: Vec<(Record, u64)> =
-                if let Some(own_buf) = drain_node_buffer_slot(ctx, node_idx) {
-                    let (records, _puncts) = own_buf.drain_split()?;
-                    records
-                } else {
-                    let predecessors: Vec<NodeIndex> = current_dag
-                        .graph
-                        .neighbors_directed(node_idx, Direction::Incoming)
-                        .collect();
-                    let mut found: Option<Vec<(Record, u64)>> = None;
-                    for &p in &predecessors {
-                        // When multiple outputs share a predecessor,
-                        // clone the buffer for all but the last
-                        // consumer to avoid starving siblings.
-                        let remaining_consumers = current_dag
-                            .graph
-                            .neighbors_directed(p, Direction::Outgoing)
-                            .filter(|&succ| succ > node_idx)
-                            .count();
-                        if remaining_consumers == 0 {
-                            if let Some(nb) = drain_node_buffer_slot(ctx, p) {
-                                let (records, _puncts) = nb.drain_split()?;
-                                found = Some(records);
-                                break;
-                            }
-                        } else {
-                            // Multi-consumer fanout: keep the producer's
-                            // buffer alive for remaining siblings via a
-                            // heap clone. The clone's footprint flows
-                            // through pull-mode attribution at the
-                            // arbitrator's next poll; the producer slot's
-                            // registered consumer keeps reporting the live
-                            // buffer until the last consumer drains it.
-                            //
-                            // Output is terminal — it writes to disk and
-                            // has no downstream node_buffer to receive
-                            // forwarded punctuations. The fan-out clone
-                            // path takes records only; sibling consumers
-                            // that need the document boundary read it from
-                            // the non-cloned (last-consumer) drain through
-                            // their own arm.
-                            let cloned = ctx.node_buffers.get(&p).map(|nb| {
-                                nb.clone_memory_only()
-                                    .into_iter()
-                                    .filter_map(|e| e.into_record())
-                                    .collect::<Vec<(Record, u64)>>()
-                            });
-                            if let Some(cloned) = cloned {
-                                found = Some(cloned);
-                                break;
-                            }
-                        }
-                    }
-                    found.unwrap_or_default()
-                };
-
-            if let Some(expected) = current_dag.graph[node_idx]
-                .expected_input_schema_in(current_dag)
-                .cloned()
-            {
-                let upstream_name = current_dag
-                    .graph
-                    .neighbors_directed(node_idx, Direction::Incoming)
-                    .next()
-                    .map(|i| current_dag.graph[i].name().to_string())
-                    .unwrap_or_default();
-                for (record, _) in &input_records {
-                    check_input_schema(&expected, record.schema(), name, "output", &upstream_name)?;
-                }
-            }
-
-            // When correlation buffering is active, every record
-            // routed to this Output goes through the per-group buffer
-            // — `CorrelationCommit` decides at end-of-DAG whether to
-            // flush the group to the writer or DLQ it. Null-keyed
-            // records get a row-disambiguated buffer cell each so
-            // they retain per-record-rejection semantics without
-            // splitting the writer path.
-            let buffered: Vec<(Record, u64, Vec<GroupByKey>)>;
-            let unbuffered: Vec<(Record, u64)>;
-            if ctx.correlation_buffers.is_some() {
-                buffered = input_records
-                    .into_iter()
-                    .map(|(rec, rn)| {
-                        let key = buffer_key_for_record(&rec, rn);
-                        (rec, rn, key)
-                    })
-                    .collect();
-                unbuffered = Vec::new();
-            } else {
-                buffered = Vec::new();
-                unbuffered = input_records;
-            }
-            // Counter semantics:
-            //
-            // * `records_written` increments per WRITE — under
-            //   inclusive Route fan-out, one input matching N
-            //   branches counts N (one per Output that received
-            //   it). Aligns with per-Output throughput and the
-            //   `records_emitted` local that drives stage-metric
-            //   reporting.
-            //
-            // * `ok_count` increments by the number of DISTINCT
-            //   source rows reaching this Output that haven't
-            //   already been counted at another Output during
-            //   the same DAG walk. Source identity is
-            //   `row_num` (per-source counter), tracked across
-            //   all Output arms via the `ok_source_rows` set
-            //   declared at function scope.
-            //
-            // Buffered records DEFER counter increments to the
-            // `CorrelationCommit` arm — clean groups bump
-            // counters at flush time; dirty groups never count
-            // toward `ok_count`.
-            let unbuffered_record_count = unbuffered.len() as u64;
-            let mut newly_ok: u64 = 0;
-            for (_, row_num) in &unbuffered {
-                if ctx.ok_source_rows.insert(*row_num) {
-                    newly_ok += 1;
-                }
-            }
-            ctx.counters.ok_count += newly_ok;
-            ctx.counters.records_written += unbuffered_record_count;
-            ctx.records_emitted += unbuffered_record_count;
-
-            // Derive output schema from first emitted record.
-            // The Record is authoritative post-rip; materialize
-            // the output-projection's `emitted` / `metadata`
-            // maps from it on demand at this boundary. That
-            // pays the bucket-insert cost once per record
-            // reaching the writer, not every intermediate node
-            // transition (Invariant 3).
-            let scan_timer = stage_metrics::StageTimer::new(stage_metrics::StageName::SchemaScan);
-            let out_cfg = ctx
-                .output_configs
-                .iter()
-                .find(|o| o.name == *name)
-                .unwrap_or(ctx.primary_output);
-
-            // Compute the upstream node's CXL emit names once.
-            // `include_unmapped: false` consults this to drop upstream
-            // passthroughs the user did not explicitly emit.
-            let cxl_emit_names = current_dag.graph[node_idx].cxl_emit_names_in(current_dag);
-            let cxl_emit_names_opt: Option<&[String]> = if cxl_emit_names.is_empty() {
-                None
-            } else {
-                Some(&cxl_emit_names)
-            };
-
-            // Buffer non-null-key records. Project once, push slot.
-            // Overflow check fires the moment a group's record count
-            // exceeds the configured cap; subsequent records of the
-            // same group are still admitted so they can become
-            // collateral entries when `CorrelationCommit` drains the
-            // group, but admission flips the overflow flag so the
-            // commit arm emits a `GroupSizeExceeded` trigger.
-            if !buffered.is_empty() {
-                let max_buf = ctx.correlation_max_group_buffer;
-                let buffers = ctx
-                    .correlation_buffers
-                    .as_mut()
-                    .expect("correlation_buffers is Some — we just checked above");
-                for (record, rn, group_key) in buffered.iter() {
-                    let projected = project_output_from_record(record, out_cfg, cxl_emit_names_opt);
-                    let entry = buffers.entry(group_key.clone()).or_default();
-                    entry.total_records += 1;
-                    if max_buf > 0 && entry.total_records > max_buf {
-                        entry.overflowed = true;
-                    }
-                    entry.records.push(CorrelationRecordSlot {
-                        row_num: *rn,
-                        original_record: record.clone(),
-                        projected,
-                        output_name: name.clone(),
-                    });
-                }
-            }
-
-            if unbuffered.is_empty() {
-                ctx.collector.record(scan_timer.finish(0, 0));
-                return Ok(());
-            }
-
-            let output_schema = {
-                let projected = {
-                    let _guard = ctx.projection_timer.guard();
-                    project_output_from_record(&unbuffered[0].0, out_cfg, cxl_emit_names_opt)
-                };
-                Arc::clone(projected.schema())
-            };
-
-            // Find and take the writer for this output. Errors from
-            // build_format_writer / write_record / flush are captured
-            // into `output_errors` instead of short-circuiting via `?`
-            // so siblings still get their chance to fail.
-            //
-            // Fan-out path: when the plan flagged this Output for
-            // per-source-file routing, each record's source_file Arc
-            // selects the right writer; the registry holds N writers
-            // (one per discovered file).
-            if let Some(per_file) = ctx.fan_out_writers.remove(name) {
-                emit_fan_out(
-                    name,
-                    out_cfg,
-                    cxl_emit_names_opt,
-                    &output_schema,
-                    &unbuffered,
-                    per_file,
-                    &mut ctx.output_errors,
-                    &mut ctx.write_timer,
-                    &mut ctx.projection_timer,
-                    ctx.collector,
-                    scan_timer,
-                );
-            } else if let Some(raw_writer) = ctx.writers.remove(name) {
-                match build_format_writer(out_cfg, raw_writer, Arc::clone(&output_schema)) {
-                    Ok(mut csv_writer) => {
-                        ctx.collector.record(scan_timer.finish(1, 1));
-                        let mut write_failed = false;
-                        for (record, _rn) in &unbuffered {
-                            let projected = {
-                                let _guard = ctx.projection_timer.guard();
-                                project_output_from_record(record, out_cfg, cxl_emit_names_opt)
-                            };
-                            let write_result = {
-                                let _guard = ctx.write_timer.guard();
-                                csv_writer.write_record(&projected)
-                            };
-                            if let Err(e) = write_result {
-                                ctx.output_errors.push(PipelineError::from(e));
-                                write_failed = true;
-                                break;
-                            }
-                        }
-                        if !write_failed {
-                            let flush_result = {
-                                let _guard = ctx.write_timer.guard();
-                                csv_writer.flush()
-                            };
-                            if let Err(e) = flush_result {
-                                ctx.output_errors.push(PipelineError::from(e));
-                            }
-                        }
-                    }
-                    Err(e) => ctx.output_errors.push(e),
-                }
-            }
+        PlanNode::Output { .. } => {
+            crate::executor::output_dispatch::dispatch_output(ctx, current_dag, node_idx, &node)?;
         }
 
-        PlanNode::Composition { ref name, body, .. } => {
-            // Recursive body execution: collect parent-scope records
-            // per declared input port, swap `current_dag` to the body's
-            // mini-DAG, walk the body's topo, then collect the body's
-            // first declared output port and write it to this node's
-            // buffer in the parent scope. The dispatcher arm logic
-            // never diverged across body and top-level walks — both
-            // run through `dispatch_plan_node` after a current_dag
-            // swap, mirroring DataFusion's `RecursiveQueryExec` pattern
-            // where the recursive term re-enters the same execution
-            // loop with a different plan.
-            debug_assert_ne!(
-                body,
-                crate::plan::composition_body::CompositionBodyId::SENTINEL,
-                "composition {name:?}: body_id is sentinel — bind_composition did not run"
-            );
-
-            // Verify the body exists; the value itself is no longer
-            // needed at this scope — schema checks read the parent
-            // graph directly and `collect_port_records` walks live
-            // edges, so the body's lookup is deferred to
-            // `execute_composition_body`.
-            ctx.artifacts
-                .body_of(body)
-                .ok_or_else(|| PipelineError::compose_body_missing(name.clone()))?;
-
-            // Schema-check parent records before stepping into the
-            // body. Failures here surface with the parent-scope
-            // upstream name, matching the diagnostic shape every
-            // other arm emits at its own entry.
-            let predecessors: Vec<NodeIndex> = current_dag
-                .graph
-                .neighbors_directed(node_idx, Direction::Incoming)
-                .collect();
-            if let Some(expected) = current_dag.graph[node_idx]
-                .expected_input_schema_in(current_dag)
-                .cloned()
-            {
-                let upstream_name = predecessors
-                    .first()
-                    .map(|&i| current_dag.graph[i].name().to_string())
-                    .unwrap_or_default();
-                // Peek-only schema check; the records are still owned
-                // by their producer's buffer until `collect_port_records`
-                // claims them below.
-                if let Some(&first_pred) = predecessors.first()
-                    && let Some(records) = ctx.node_buffers.get(&first_pred)
-                {
-                    for (record, _) in records.peek_mem_records() {
-                        check_input_schema(
-                            &expected,
-                            record.schema(),
-                            name,
-                            "composition",
-                            &upstream_name,
-                        )?;
-                    }
-                }
-            }
-
-            // Depth guard before recursion — same constant the
-            // compile-time IsolatedFromAbove check uses, distinct
-            // emission code for log greppability.
-            if ctx.recursion_depth >= crate::plan::bind_schema::MAX_COMPOSITION_DEPTH {
-                return Err(PipelineError::compose_depth_exceeded(
-                    name.clone(),
-                    ctx.recursion_depth,
-                ));
-            }
-
-            let composition_name = name.clone();
-            let port_records = collect_port_records(ctx, current_dag, node_idx, &composition_name)?;
-            // Direct sync recursion into the body executor. Depth is
-            // bounded by MAX_COMPOSITION_DEPTH (compile-time) plus the
-            // E112 runtime cap, so the recursion cannot overflow the
-            // stack.
-            let output_records =
-                execute_composition_body(ctx, body, port_records, &composition_name)?;
-            // Materialize node-rooted window runtimes for any IndexSpec
-            // rooted at this composition's call-site NodeIndex. The
-            // body executor returned with `active_stack` already
-            // popped, so the install lands on `top` (parent scope).
-            finalize_node_rooted_windows(ctx, current_dag, node_idx, &output_records)?;
-            tee_emit_to_region_input_buffers(ctx, current_dag, node_idx, &output_records)?;
-            // Attribution rule: the body harvest's bytes are charged
-            // under the parent composition's name so a budget-exceeded
-            // diagnostic points the user at the user-visible operator,
-            // not at the body's internal output port node name.
-            //
-            // Boundary admit (post-body): same two-path model as the
-            // pre-body port-records clone in `collect_port_records`.
-            // If `admit_node_buffer` returns `MemoryBudgetExceeded`
-            // here, it surfaces bare with `node = composition_name` —
-            // not wrapped in `CompositionBodyError`. The wrapper at
-            // `execute_composition_body`'s topo walk has already
-            // returned with `Ok` by the time we reach this admit;
-            // only errors from inside that walk get the wrapper.
-            let nb = admit_node_buffer(
+        PlanNode::Composition { .. } => {
+            crate::executor::composition_dispatch::dispatch_composition(
                 ctx,
-                &composition_name,
+                current_dag,
                 node_idx,
-                output_records,
-                Vec::new(),
-                node_buffer_spill_allowed(current_dag, node_idx),
+                &node,
             )?;
-            ctx.node_buffers.insert(node_idx, nb);
         }
 
         PlanNode::Combine { .. } => {
@@ -3397,288 +2501,4 @@ pub(crate) struct CorrelationErrorRecord {
     pub(crate) error_message: String,
     pub(crate) stage: Option<String>,
     pub(crate) route: Option<String>,
-}
-
-/// Collect parent-scope records keyed by composition input port name.
-///
-/// Resolves ports via the live edge graph: walks `parent_dag`'s incoming
-/// edges into `composition_node_idx`, reads each edge's `port` tag, and
-/// clones records from the producer's `node_buffers` slot. The frozen
-/// port-name snapshot kept by an earlier design drifted whenever a
-/// planner pass spliced a node between the producer and the composition
-/// (the synthetic `inject_correlation_sort` Sort being the canonical
-/// trigger), silently emptying the producer's buffer one hop downstream.
-/// Edge-walking the live graph is the same source-of-truth pattern every
-/// other arm of the dispatcher uses (`Transform`, `Aggregate`, `Output`,
-/// `Sort` all read predecessors via `neighbors_directed`).
-///
-/// Cloning rather than removing keeps the parent producer's buffer
-/// intact for any sibling consumer the parent walk has not yet reached;
-/// fan-out from a single producer to multiple ports is a normal case.
-fn collect_port_records(
-    ctx: &mut ExecutorContext<'_>,
-    parent_dag: &ExecutionPlanDag,
-    composition_node_idx: NodeIndex,
-    composition_name: &str,
-) -> Result<IndexMap<String, Vec<(clinker_record::Record, u64)>>, PipelineError> {
-    let mut result: IndexMap<String, Vec<(clinker_record::Record, u64)>> = IndexMap::new();
-    use petgraph::visit::EdgeRef;
-    for edge in parent_dag
-        .graph
-        .edges_directed(composition_node_idx, Direction::Incoming)
-    {
-        let Some(port_name) = edge.weight().port.as_ref() else {
-            // Composition incoming edges are always port-tagged at
-            // bind time; an untagged edge is a planner-pass bug
-            // (likely a rewrite that forgot to preserve the tag) and
-            // surfaces here as an internal error rather than a silent
-            // record drop.
-            return Err(PipelineError::Internal {
-                op: "composition",
-                node: composition_name.to_string(),
-                detail: format!(
-                    "untagged incoming edge from node {:?}; every composition input edge must \
-                     carry a port name (planner-pass invariant — see PlanEdge.port)",
-                    parent_dag.graph[edge.source()].name(),
-                ),
-            });
-        };
-        // Compute the heap-clone footprint while the immutable borrow
-        // of `node_buffers` is live, then release it so the charge
-        // path can take `&mut ctx.memory_budget`. The cloned records
-        // are about to be seeded into the body-local `node_buffers`
-        // namespace inside `execute_composition_body`; the body's
-        // port-source Source arm discharges the slot when it claims
-        // the seeded buffer, and the body-exit leak-discharge cleans
-        // up any unconsumed remainder if a body operator errored.
-        //
-        // Boundary admit: the budget exceedance surfaces as a bare
-        // `MemoryBudgetExceeded` with `node = composition_name`
-        // (the user-visible call-site identifier). The body has not
-        // executed yet, so there is no body-internal operator name
-        // to disambiguate via `CompositionBodyError` — only inner
-        // errors that bubble out of `execute_composition_body`'s
-        // topo walk get that wrapper. Detail string discriminates
-        // this site from the generic admit in `admit_node_buffer`.
-        let cloned_with_bytes = ctx.node_buffers.get(&edge.source()).map(|nb| {
-            // Composition port seeding clones records only; the
-            // composition body operates inside its own document-
-            // boundary scope and re-emits punctuations at the call-
-            // site level on body exit, so the parent's puncts do not
-            // forward through the body's port-source boundary.
-            let cloned_records: Vec<(Record, u64)> = nb
-                .clone_memory_only()
-                .into_iter()
-                .filter_map(|e| e.into_record())
-                .collect();
-            (cloned_records, nb.estimated_memory_bytes())
-        });
-        let records: Vec<(Record, u64)> = match cloned_with_bytes {
-            Some((cloned, _bytes)) => cloned,
-            None => Vec::new(),
-        };
-        // Two parallel edges to the same port (e.g. `inputs: { p: a,
-        // p: a }` — currently rejected at parse, but the runtime is
-        // defensive) would overwrite; the wiring pass guarantees
-        // unique port names per consumer.
-        result.insert(port_name.clone(), records);
-    }
-    Ok(result)
-}
-
-/// Execute one composition body's mini-DAG.
-///
-/// Builds a transient body-scope `ExecutionPlanDag` and walks it
-/// through `dispatch_plan_node` — the same dispatcher entry the
-/// top-level walker uses. The body's `node_buffers` namespace
-/// is swapped in via `mem::replace` so body NodeIndices index a
-/// fresh space; the parent buffers are restored after the walk.
-/// The depth-counter guard increments via RAII so `?`-bubbled
-/// errors can't leak the counter.
-fn execute_composition_body(
-    ctx: &mut ExecutorContext<'_>,
-    body_id: crate::plan::composition_body::CompositionBodyId,
-    port_records: IndexMap<String, Vec<(clinker_record::Record, u64)>>,
-    composition_name: &str,
-) -> Result<Vec<(clinker_record::Record, u64)>, PipelineError> {
-    use crate::plan::index::PlanIndexRoot;
-
-    // Resolve body and pre-compute everything that needs the
-    // bound_body borrow before the swap so the body_dag clone is
-    // independent of the artifacts borrow.
-    let bound_body = ctx
-        .artifacts
-        .body_of(body_id)
-        .ok_or_else(|| PipelineError::compose_body_missing(composition_name.to_string()))?;
-
-    let body_dag = crate::plan::execution::ExecutionPlanDag::from_body(bound_body);
-
-    // Window runtime entry: install a fresh per-body vec sized to
-    // `body_indices_to_build.len()`. ParentNode-rooted slots inherit
-    // the parent's runtime via `Arc::clone` from `top` so the body's
-    // windowed Transform arm can resolve through to it without re-
-    // materializing the parent's arena. Node-rooted body slots stay
-    // `None` here — they populate when the body's upstream operator
-    // arm calls `finalize_node_rooted_windows` during the body walk.
-    let body_index_count = bound_body.body_indices_to_build.len();
-    let mut body_window_vec: Vec<Option<crate::executor::window_runtime::WindowRuntime>> =
-        (0..body_index_count).map(|_| None).collect();
-    for (idx, spec) in bound_body.body_indices_to_build.iter().enumerate() {
-        if matches!(spec.root, PlanIndexRoot::ParentNode { .. })
-            && let Some(parent_runtime) = ctx.window_runtime.resolve(&spec.root, idx)
-        {
-            // `resolve` for ParentNode in this position falls through
-            // to top[idx] (no body is on the active_stack yet), which
-            // is correct: the body inherits the parent operator's
-            // runtime by cloning its `Arc<Arena>` and `Arc<SecondaryIndex>`.
-            body_window_vec[idx] = Some(parent_runtime.clone());
-        }
-    }
-
-    // Seed body-scope buffers from parent records keyed by port.
-    let mut body_buffers: HashMap<NodeIndex, NodeBuffer> = HashMap::new();
-    for (port_name, records) in port_records {
-        let body_idx = bound_body
-            .port_name_to_node_idx
-            .get(port_name.as_str())
-            .ok_or_else(|| PipelineError::compose_unknown_port(composition_name, &port_name))?;
-        body_buffers.insert(*body_idx, NodeBuffer::memory_from_records(records));
-    }
-
-    // Pick the body's terminal output node. The bind-time alias
-    // resolution wrote the port → NodeIndex map onto BoundBody;
-    // the first declared output port wins. Zero-output-port bodies
-    // are legal (sink-only / side-effect bodies) and produce no
-    // record stream back to the parent.
-    let output_idx = bound_body.output_port_to_node_idx.values().next().copied();
-
-    // Swap node_buffers to a body-local namespace so body NodeIndices
-    // don't collide with the parent's. `source_records` is also
-    // swapped to an empty map so body-scope Source nodes resolve
-    // through `node_buffers` (port seeding from parent scope), not
-    // through parent-scope source ingestion — bodies declare ports,
-    // not top-level sources. Any non-port-seeded body Source surfaces
-    // as the defense-in-depth `Internal` error from the Source arm.
-    let saved_buffers = std::mem::replace(&mut ctx.node_buffers, body_buffers);
-    let saved_combine = std::mem::take(&mut ctx.source_records);
-    // Window-arena consumer ids key by slot index, which the body
-    // re-uses from zero alongside its window-runtime overlay. Swap to a
-    // fresh map so a body arena registration at slot N does not clobber
-    // the parent's slot N; body-local arenas drop when the body's
-    // window-runtime overlay is popped on exit, so their wrappers are
-    // unregistered there and the parent map is restored.
-    let saved_arena_ids = std::mem::take(&mut ctx.window_arena_consumer_ids);
-    // Install the body's `input:` reference table so the Route arm
-    // can resolve `<route>.<branch>` references against body
-    // siblings. Restored on exit.
-    let saved_body_refs = ctx
-        .current_body_node_input_refs
-        .replace(bound_body.node_input_refs.clone());
-
-    // Push the body's window-runtime overlay onto the registry. The
-    // ParentNode-rooted slots were populated above via `Arc::clone`
-    // from the parent's `top` runtime; Node-rooted body slots stay
-    // `None` and populate when the body's upstream operator arm
-    // calls `finalize_node_rooted_windows` during the body walk.
-    // `active_stack.push` makes `resolve` route through this overlay
-    // for any window dispatched inside the body.
-    ctx.window_runtime.bodies.insert(body_id, body_window_vec);
-    ctx.window_runtime.active_stack.push(body_id);
-
-    // Increment depth before recursing. Every exit path below
-    // decrements before returning so the counter stays in sync —
-    // including the `walk_result?` early-return at the end. The
-    // dispatcher loop already collects errors into `walk_result`
-    // rather than `?`-bubbling, so the only `?`-bubble that escapes
-    // this function is on `walk_result?` itself, which fires AFTER
-    // the decrement.
-    ctx.recursion_depth += 1;
-
-    // Walk the body's topo through the same dispatcher the top-level
-    // walker uses. Errors from within the body are wrapped with the
-    // composition's name for diagnosability — the user sees
-    // "in composition '<name>': <inner>" instead of an opaque
-    // inner-only message.
-    //
-    // Body-interior path of the two-path admission model: the inner
-    // error's `node` field names a body-internal operator (e.g.
-    // `stage_split`) the user never wrote in their YAML, so the
-    // wrapper supplies the user-visible call-site name on top. The
-    // boundary admits in `collect_port_records` (pre-body input
-    // clone) and the parent's Composition arm (post-body output
-    // harvest) stay unwrapped because their `node` is already the
-    // call-site composition name — wrapping them would duplicate
-    // the same identifier in both layers with no information gain.
-    // Consumers that want to catch every composition-involved
-    // budget exceedance must match both shapes.
-    let topo: Vec<NodeIndex> = body_dag.topo_order.clone();
-    let mut walk_result: Result<(), PipelineError> = Ok(());
-    for node_idx in topo {
-        if let Err(inner) = dispatch_plan_node(ctx, &body_dag, node_idx) {
-            walk_result = Err(PipelineError::compose_body_error(
-                composition_name.to_string(),
-                Box::new(inner),
-            ));
-            break;
-        }
-    }
-
-    // Harvest output before restoring parent buffers. When the output
-    // port aliases a deferred-region producer (a relaxed-CK Aggregate
-    // sitting at the body's terminal port), the producer's forward
-    // emit is NOT a final stream — the commit-pass body→parent
-    // harvest path re-emits the post-recompute narrow rows into the
-    // parent's `node_buffers[composition_idx]`. Returning the forward
-    // emit here would double-feed the parent's continuation: once
-    // through the parent Composition arm's `node_buffers.insert`, then
-    // again when `recurse_into_body` extends the slot with the
-    // commit-pass harvest. Drop the forward emit so the commit pass
-    // is the single source of records for that slot.
-    let output_records: Vec<(Record, u64)> = match (&walk_result, output_idx) {
-        (Ok(()), Some(idx)) => {
-            // Composition body output harvest — punctuations on the
-            // body's output port belong to the composition's parent
-            // pipeline scope; they re-emit when the parent's call
-            // site re-introduces document context. The body harvest
-            // takes records only.
-            let drained: Vec<(Record, u64)> = match drain_node_buffer_slot(ctx, idx) {
-                Some(nb) => {
-                    let (records, _puncts) = nb.drain_split()?;
-                    records
-                }
-                None => Vec::new(),
-            };
-            if body_dag.deferred_region_at_producer(idx).is_some() {
-                Vec::new()
-            } else {
-                drained
-            }
-        }
-        _ => Vec::new(),
-    };
-
-    // Decrement depth and restore parent scope. `saturating_sub`
-    // is defensive over the invariant; the inc/dec pairs are kept in
-    // sync by hand on every exit path through this function.
-    ctx.recursion_depth = ctx.recursion_depth.saturating_sub(1);
-    ctx.node_buffers = saved_buffers;
-    ctx.source_records = saved_combine;
-    ctx.current_body_node_input_refs = saved_body_refs;
-    // Unregister body-local window-arena consumers and restore the
-    // parent map. The body's node-rooted arenas drop when its window-
-    // runtime overlay is popped below, so their wrappers must leave the
-    // arbitrator's registry in lockstep.
-    for (_, (id, _)) in std::mem::take(&mut ctx.window_arena_consumer_ids) {
-        ctx.memory_budget.unregister_consumer(id);
-    }
-    ctx.window_arena_consumer_ids = saved_arena_ids;
-    // Pop the window-runtime overlay so subsequent windows in the
-    // parent scope route through `top` again. Removing the body's
-    // entry releases the `Arc` clones (parent runtimes stay alive in
-    // `top`; body-local node-rooted runtimes drop here).
-    ctx.window_runtime.active_stack.pop();
-    ctx.window_runtime.bodies.remove(&body_id);
-
-    walk_result?;
-    Ok(output_records)
 }
